Added in Player Jump and Implemented Base for Jump Pad.
For now Player Jump is done. Need a bit of tweaking since the Physics.gravity has to be adjusted in order to cater to best user experience as the default gravity has too slow of a fall speed. Jump Pad acts as a teleporter for now. Need to do the Arc to get it working.