Edited IGun Script and IProjectile Script. Also made the Rocket Launcher, Shockwave Gun and Shotgun.
Did some base testing. Everything seems to be working. Raycast Shotgun does seem to have a very poor accuracy for now and Projectile Shotgun seems to be OP for the moment as it seems to be able to travel quite a distance. However, base functionalities as a concept is there.